Improve the performance of prom_info.metric
The prom_info.metric view can be slow when there are many chunks. The original query uses a number of functions/views. By "inlining" these and modifying them to be less generic (e.g. we KNOW that metrics only have a single dimension), we can speed things up. Furthermore, calling pg_total_relation_size is really slow relative to the rest of the query. We don't need to call pg_total_relation_size for compressed chunks. The size of compressed chunks should be static and is already recorded in a table.